created density.py
